Fix delete not checking for table existence

+++ GangsImpl/SQL/SQLInstanceManager.cs
@@ -72,6 +72,9 @@ public class SQLInstanceManager(string connectionString, string table_prefix,
   }
 
   public async Task<bool> RemoveFromGang(int gangId, string statId) {
+    var result = await connection.QuerySingleOrDefaultAsync(
+      $"SELECT 1 FROM information_schema.tables WHERE table_name = '{table_prefix}_{statId}'");
+    if (result == null) return false;
     return await connection.ExecuteAsync(
       $"DELETE FROM {table_prefix}_{statId} WHERE GangId = @GangId",
       new { GangId = gangId }) > 0;
